+++ scratch.js
@@ -2,37 +2,40 @@
 const wtf = require('./src/index');
 // const wtf = require('./builds/wtf_wikipedia');
 // const wtf = require('./build');
-// let parse = wtf.parse;
-
-// wtf.from_api("Tomb_Raider_(2013_video_game)", 'en', function(s) {
-//   console.log(wtf.parse(s).infobox)
-// })
-//
-// wtf.from_api('Raoul Dautry', 'fr',function(page) {
-//   var parsed = wtf.parse(page); // causes the crash
-//   console.log(parsed.text.get('Intro'));
-// });
-
 function from_file(page) {
   let str = require('fs').readFileSync('./tests/cache/' + page.toLowerCase() + '.txt', 'utf-8');
-  console.log(wtf.plaintext(str));
-  // let r = wtf.parse(str, {});
-  // console.log(r);
+  // console.log(wtf.plaintext(str));
+  let r = wtf.parse(str);
+  console.log(r.translations);
   // console.log(JSON.stringify(r.sections, null, 2));
 }
 
 // from_file("list")
 // from_file("earthquakes");
 // from_file("bluejays");
-// from_file("Toronto")
+// from_file('Toronto');
 // from_file('Toronto_Star');
-from_file('royal_cinema');
+// from_file('Radiohead');
 // from_file('Jodie_Emery');
 // from_file('Redirect')
 // from_file("Africaans")
 // from_file('Anarchism');
 
-// wtf.from_api('Paris', 'en', function(markup) {
+// wtf.from_api('Toronto', 'en', function(markup) {
 //   var obj = wtf.parse(markup);
 //   console.log(obj);
 // });
+
+// {
+//   type:'', //'page'/redirect/category/template
+//   infoboxes:[], //parsed key-value for each infobox template
+//   categories:[], //list of parsed categories
+//   interwiki:{},
+//   sections:[{
+//       title:'',
+//       images:'',
+//       lists:'',
+//     }]
+// }
+var str = 'hello [[wikinews:Radiohead]] world  [[Category:Films]]';
+console.log(wtf.parse(str));