+++ changelog.md
@@ -4,3 +4,5 @@
 ## 2.0.0
 * turn `infobox` into 'infoboxes' array
 * remove top-level `infobox_template`
+* change initial depth to 0
+* change 'translations' to 'interwiki'